requirements.txt
Django==3.2.13
django-heroku==0.3.1
djangorestframework==3.13.1
django-rest-auth==0.9.5
gunicorn==20.0.4
requests==2.25.1